requirements.txt
albumentations==1.4.21
matplotlib==3.8.2
numpy==2.2.3
opencv_contrib_python==3.4.13.47
opencv_python==3.4.13.47
opencv_python_headless==3.4.13.47
Pillow==9.5.0
Pillow==11.1.0
torch==1.8.2+cu111
torchvision==0.9.2+cu111
tqdm==4.66.4
